Move all rekor code to tlog
The code in `signing.go` makes little sense, why not have it all
in tlog. I made this change while looking at leveraging the code
within the tektoncd/chains project and importing sigstore/signfile
looks confusing.
