+++ cmd/sign.go
@@ -20,6 +20,7 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
+	"github.com/sigstore/sigstore/pkg/tlog"
 	"io/ioutil"
 	"net/http"
 	"os"
@@ -31,7 +32,6 @@ import (
 	"github.com/sigstore/sigstore/pkg/httpclients"
 	"github.com/sigstore/sigstore/pkg/oauthflow"
 	"github.com/sigstore/sigstore/pkg/signature"
-	"github.com/sigstore/sigstore/pkg/signfile"
 	"github.com/sigstore/sigstore/pkg/utils"
 )
 
@@ -134,7 +134,7 @@ var signCmd = &cobra.Command{
 
 		// Send to rekor
 		fmt.Println("Sending entry to transparency log")
-		tlogEntry, err := signfile.UploadToRekor(
+		tlogEntry, err := tlog.UploadToRekor(
 			pub,
 			signedVal,
 			signature,
